+++ packages/progress-indicator/src/js/react.js
@@ -7,7 +7,7 @@
  *
  **************************************************************************************************************************************************************/
 
-import React, { Fragment } from 'react';
+import React from 'react';
 import PropTypes from 'prop-types';
 
 
@@ -32,6 +32,7 @@ const StatusText = {
  * An item inside the AUprogressIndicator component
  *
  * @param  {string}   link             - The link URL, If no link is passed we render a button instead of a link tag, optional
+ * @param  {string}  linkComponent     - The component used for the link
  * @param  {string}   text             - The text of this item
  * @param  {string}   status           - The status of this item
  * @param  {string}   statusText       - The status text of this item, optional
@@ -39,21 +40,38 @@ const StatusText = {
  * @param  {object}   li               - An additional object to be spread into the wrapping element, optional
  * @param  {object}   attributeOptions - Any other attribute options
  */
-export const AUprogressIndicatorItem = ({ link, text, status, statusText, className = '', li = {}, ...attributeOptions }) => (
-	<li { ...li }>
+export const AUprogressIndicatorItem = ({ link, linkComponent, text, status, statusText, className = '', li = {}, ...attributeOptions }) => {
+	const LinkComponent = linkComponent;
+
+	if( link ){
+		// If we are using a normal link
+		if( LinkComponent === 'a' ) {
+			attributeOptions.href = link;
+		}
+		// If we are using a link component
+		else if( typeof LinkComponent === 'function' ) {
+			attributeOptions.to = link;
+		}
+	}
+
+	return(
+		<li { ...li }>
 		{
 			link
-				? <a className={`au-progress-indicator__link au-progress-indicator__link--${ status } ${ className }`} href={ link } { ...attributeOptions }>
+				?
+					<LinkComponent className={`au-progress-indicator__link au-progress-indicator__link--${ status } ${ className }`} { ...attributeOptions }>
 						<span className="au-progress-indicator__status">{ statusText ? statusText : StatusText[ status ] }</span>
 						{ text }
-					</a>
-				: <button className={`au-progress-indicator__link au-progress-indicator__link--${ status } ${ className }`} { ...attributeOptions }>
+					</LinkComponent>
+				:
+					<button className={`au-progress-indicator__link au-progress-indicator__link--${ status } ${ className }`} { ...attributeOptions }>
 						<span className="au-progress-indicator__status">{ statusText ? statusText : StatusText[ status ] }</span>
 						{ text }
 					</button>
 		}
 	</li>
-);
+	)
+};
 
 AUprogressIndicatorItem.propTypes = {
 	link: PropTypes.string,
@@ -61,6 +79,12 @@ AUprogressIndicatorItem.propTypes = {
 	status: PropTypes.PropTypes.oneOf([ 'doing', 'todo', 'done' ]).isRequired,
 	statusText: PropTypes.string,
 	li: PropTypes.object,
+	linkComponent: PropTypes.oneOfType([ PropTypes.string, PropTypes.func ]),
+};
+
+
+AUprogressIndicatorItem.defaultProps = {
+	linkComponent: 'a',
 };
 
 
@@ -71,13 +95,14 @@ AUprogressIndicatorItem.propTypes = {
  * @param  {boolean} dark             - Add the dark variation class, optional
  * @param  {array}   items            - All items for this progress indicator
  * @param  {string}  className        - An additional class, optional
+ * @param  {string}  linkComponent    - The component used for the link
  * @param  {object}  attributeOptions - Any other attribute options
  */
-const AUprogressIndicator = ({ dark, items, className = '', ...attributeOptions }) => (
+const AUprogressIndicator = ({ dark, linkComponent, items, className = '', ...attributeOptions }) => (
 	<ul className={ `au-progress-indicator${ dark ? ' au-progress-indicator--dark' : '' } ${ className }` } { ...attributeOptions }>
 		{
 			items.map(
-				( item, i ) => <AUprogressIndicatorItem key={ i } { ...item } />
+				( item, i ) => <AUprogressIndicatorItem linkComponent={ linkComponent } key={ i } { ...item } />
 			)
 		}
 	</ul>
@@ -94,6 +119,11 @@ AUprogressIndicator.propTypes = {
 			li: PropTypes.object,
 		})
 	).isRequired,
+	linkComponent: PropTypes.oneOfType([ PropTypes.string, PropTypes.func ]),
+};
+
+AUprogressIndicator.defaultProps = {
+	linkComponent: 'a',
 };
 
 export default AUprogressIndicator;