+++ packages/tags/tests/react/index.js
@@ -1,5 +1,6 @@
-import React from 'react';
+import React, { Fragment } from 'react';
 import ReactDOM from 'react-dom';
+import { BrowserRouter, Route, Link } from 'react-router-dom';
 
 import AUtags from './tags.js';
 
@@ -171,6 +172,27 @@ ReactDOM.render(
 					onClick: event => { event.preventDefault(); console.log('This function is called when the forth tag is clicked') },
 				},
 			]} />
+
+			<hr />
+			<h2>tags with router</h2>
+
+			<BrowserRouter basename={ window.location.pathname }>
+				<Fragment>
+					<AUtags dark linkComponent={ Link } tags={[
+						{
+							link: 'one',
+							text: 'one',
+						},
+						{
+							link: 'two',
+							text: 'two',
+						}
+					]} />
+					<br />
+					<Route path="/one" render={ () => ( <p>Route one</p> )} />
+					<Route path="/two" render={ () => ( <p>Route two</p> )} />
+				</Fragment>
+			</BrowserRouter>
 		</div>
 	</div>,