+++ packages/accordion/CHANGELOG.md
@@ -15,6 +15,7 @@
 
 ## Versions
 
+* [v3.0.4 - Change npm run watch browser-sync location](#v304)
 * [v3.0.3 - au-body--dark now changes accordion title colour](#v303)
 * [v3.0.2 - Update dependencies](#v302)
 * [v3.0.1 - React accordion open by default](#v301)
@@ -34,6 +35,11 @@
 
 
 ## Release History
+
+### v3.0.4
+
+- Fix the serve script for npm
+
 
 ### v3.0.3