+++ packages/grid-12/CHANGELOG.md
@@ -15,6 +15,7 @@
 
 ## Versions
 
+* [v2.0.4 - Change npm run watch browser-sync location](#v204)
 * [v2.0.3 - Update dependencies](#v203)
 * [v2.0.2 - Change homepage link](#v202)
 * [v2.0.1 - Fix dependencies](#v201)
@@ -28,6 +29,11 @@
 
 
 ## Release History
+
+### v2.0.4
+
+- Change npm run watch browser-sync location
+
 
 ### v2.0.3